+++ service/main.go
@@ -5,6 +5,9 @@ import (
 	"net/http"
 	"os"
 
+	acoEnv "github.com/argoproj-labs/argo-cloudops/internal/env"
+	"github.com/argoproj-labs/argo-cloudops/service/internal/workflow"
+	"github.com/argoproj/argo-workflows/v3/cmd/argo/commands/client"
 	"github.com/go-kit/kit/log"
 	"github.com/go-kit/kit/log/level"
 	vault "github.com/hashicorp/vault/api"
@@ -19,12 +22,12 @@ func main() {
 		argoAddr    = os.Getenv("ARGO_ADDR")
 		sshPemFile  = os.Getenv("SSH_PEM_FILE")
 		logLevel    = os.Getenv("ARGO_CLOUD_OPS_LOG_LEVEL")
-		port        = getenv("ARGO_CLOUD_OPS_PORT", "8443")
+		port        = acoEnv.Getenv("ARGO_CLOUD_OPS_PORT", "8443")
 	)
 
 	setLogLevel(&logger, logLevel)
 
-	if len(adminSecret()) < 16 {
+	if len(acoEnv.AdminSecret()) < 16 {
 		panic("ARGO_CLOUDOPS_ADMIN_SECRET must be 16 characers long.")
 	}
 
@@ -48,12 +51,12 @@ func main() {
 		panic("SSH_PEM_FILE is undefined")
 	}
 
-	level.Info(logger).Log("message", fmt.Sprintf("loading config '%s'", configFilePath()))
+	level.Info(logger).Log("message", fmt.Sprintf("loading config '%s'", acoEnv.ConfigFilePath()))
 	config, err := loadConfig()
 	if err != nil {
 		panic(fmt.Sprintf("Unable to load config %s", err))
 	}
-	level.Info(logger).Log("message", fmt.Sprintf("loading config '%s' completed", configFilePath()))
+	level.Info(logger).Log("message", fmt.Sprintf("loading config '%s' completed", acoEnv.ConfigFilePath()))
 
 	vaultSvc, err := newVaultSvc(vaultAddr, vaultRole, vaultSecret)
 	if err != nil {
@@ -67,10 +70,18 @@ func main() {
 		panic("error creating git client")
 	}
 
+	// The Argo context is needed for any Argo client method calls or else, nil errors.
+	argoCtx, argoClient := client.NewAPIClient()
+	namespace := acoEnv.ArgoNamespace()
+
+	// Any Argo Workflow client method calls need the context returned from NewAPIClient, otherwise
+	// nil errors will occur. Mux sets its params in context, so passing the Argo Workflow context to
+	// setupRouter and applying it to the request will wipe out Mux vars (or any other data Mux sets in its context).
 	h := handler{
 		logger:                 logger,
 		newCredentialsProvider: newVaultProvider(vaultSvc),
-		argo:                   newArgoWorkflow(),
+		argo:                   workflow.NewArgoWorkflow(argoClient.NewWorkflowServiceClient(), namespace),
+		argoCtx:                argoCtx,
 		config:                 config,
 		gitClient:              gitClient,
 	}
@@ -83,8 +94,6 @@ func main() {
 		level.Error(logger).Log("message", "error starting service", "error", err)
 		panic("error starting service")
 	}
-
-	return
 }
 
 func setLogLevel(logger *log.Logger, logLevel string) {