+++ CHANGELOG.md
@@ -1,11 +1,28 @@
 # Changelog
 
-## Unreleased
+## v0.0.4 (2021-06-03)
+### Fixed
+* Passing Argo context to Argo Workflow calls
+
+## v0.0.3 (2021-05-27)
+### Changed
+* Update environmental variable name to specify workflow execution namespace
+
+### Added
 * Adds operations API for git based sync/diff
+
+### Fixed
 * Fix typos in CLI command description
+* Use GitHub Action for linting.
+* Fix issues reported by linter.
+* Perform deeper linting.
+* Add build caching.
+* Remove vendoring.
 
 ## v0.0.2 (2021-05-06)
+### Changed
 * Set service port via environment variable, default 8443
 
 ## v0.0.1 (2021-05-06)
+### Added
 * Initial release